pyproject.toml
[tool.poetry]
name = "snakes-in-a-hurry"
version = "0.1.0"
description = ""
authors = ["JirkaKlimes <jiriklimes920@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
cupy-cuda12x = "^13.1.0"
opencv-python = "^4.9.0.80"
numba = "^0.59.1"
tabulate = "^0.9.0"
jaxlib = { version = "^0.4.26+cuda12.cudnn89", platform = "linux", source = "jax" }
flax = "^0.8.2"
tqdm = "^4.66.2"

[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
priority = "primary"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
